Use the correct type to offset and size.
Size were 32 bits and offset were 64 bits.
So we were using the offset_type everywhere we need a 64 bits integer even
if it was a size.

Now, size_type is a 64 bits, reuse the correct type everywhere.
We also use internal type `offset_t` and `zsize_t` to avoid implicit
conversion (and bugs).

As size_type change from 32 bits to 64 bits, we need to read a real
`uint32_t` in the zim when necessary.

+++ include/zim/zim.h
@@ -33,7 +33,8 @@ namespace zim
   // An index of a blog (in a cluster)
   typedef uint32_t blob_index_type;
 
-  typedef uint32_t size_type;
+  // The size of something (article, zim, cluster, blob, ...)
+  typedef uint64_t size_type;
   
   // An offset.
   typedef uint64_t offset_type;
